Refactor TypeIdentifier to avoid circular imports
